Worker: extract rabbi static data + resolver into rabbi-places.ts
Second Batch-3 slice. RABBI_PLACES was defined inline in index.ts but used in
15+ places across routes, the rabbi mark path, and the daf payload — the
cross-cutting keystone for any rabbi route extraction (same role Bindings played
for the rest). Lifts the data + lookup layer into a neutral module:

- Movement, RabbiPlacesEntry, RabbiPlacesFile, RABBI_PLACES (+ the json import)
- RabbiResolution + normalizeHeForResolve / BY_CANONICAL_HE / resolveRabbiByHe /
  resolveRabbiByName / resolveRabbi / resolveRabbiName

The identity/enrichment layer (enrichRabbi, resolveGeneration, enrichAll,
isRabbinicEntry) stays in index.ts for a later slice. Pure move; rabbi-places.ts
imports only the json, never ./index (no cycle). tests/rabbi-resolver.test.ts now
imports the resolvers from their new home. index.ts 7,872 -> 7,720.
